+++ src/Stripe.net/Entities/PaymentIntents/PaymentIntentNextActionDisplayBankTransferInstructions.cs
@@ -27,6 +27,12 @@ namespace Stripe
         public List<PaymentIntentNextActionDisplayBankTransferInstructionsFinancialAddress> FinancialAddresses { get; set; }
 
         /// <summary>
+        /// A link to a hosted page that guides your customer through completing the transfer.
+        /// </summary>
+        [JsonProperty("hosted_instructions_url")]
+        public string HostedInstructionsUrl { get; set; }
+
+        /// <summary>
         /// A string identifying this payment. Instruct your customer to include this code in the
         /// reference or memo field of their bank transfer.
         /// </summary>