+++ src/Stripe.net/Services/Accounts/AccountIndividualOptions.cs
@@ -77,6 +77,17 @@ namespace Stripe
         public string IdNumber { get; set; }
 
         /// <summary>
+        /// The government-issued secondary ID number of the individual, as appropriate for the
+        /// representative's country, will be used for enhanced verification checks. In Thailand,
+        /// this would be the laser code found on the back of an ID card. Instead of the number
+        /// itself, you can also provide a <a
+        /// href="https://stripe.com/docs/js/tokens_sources/create_token?type=pii">PII token created
+        /// with Stripe.js</a>.
+        /// </summary>
+        [JsonProperty("id_number_secondary")]
+        public string IdNumberSecondary { get; set; }
+
+        /// <summary>
         /// The individual's last name.
         /// </summary>
         [JsonProperty("last_name")]